% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_deriv_dgamma_r}
\alias{deriv_deriv_dgamma_r}
\title{2nd Deriv of gamma density wrt. theta, evaluated numerically}
\usage{
deriv_deriv_dgamma_r(x, theta)
}
\description{
2nd Deriv of gamma density wrt. theta, evaluated numerically
}

