% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_deriv_dlognormal_r_numeric}
\alias{deriv_deriv_dlognormal_r_numeric}
\title{2nd deriv log-normal, numerical}
\usage{
deriv_deriv_dlognormal_r_numeric(x, theta)
}
\description{
2nd deriv log-normal, numerical
}

