% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{deriv_deriv_lt_dgamma_r}
\alias{deriv_deriv_lt_dgamma_r}
\title{Gamma LT 2nd derivative wrt. theta. This obviously was not solved by hand.}
\usage{
deriv_deriv_lt_dgamma_r(p, s, theta)
}
\description{
Gamma LT 2nd derivative wrt. theta. This obviously was not solved by hand.
}

