% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_deriv_lt_dgamma_r_numeric}
\alias{deriv_deriv_lt_dgamma_r_numeric}
\title{2nd deriv evaluated numerically}
\usage{
deriv_deriv_lt_dgamma_r_numeric(p, s, theta)
}
\description{
2nd deriv evaluated numerically
}

