% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_deriv_lt_dpvf_coef_r}
\alias{deriv_deriv_lt_dpvf_coef_r}
\title{Coefficients for the PVF LT}
\usage{
deriv_deriv_lt_dpvf_coef_r(m, j, alpha)
}
\description{
Coefficients for the PVF LT
}

