% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{deriv_deriv_lt_dpvf_coef_r_numeric}
\alias{deriv_deriv_lt_dpvf_coef_r_numeric}
\title{Coefficients for the PVF LT, numerical deriv wrt. alpha}
\usage{
deriv_deriv_lt_dpvf_coef_r_numeric(p, j, alpha)
}
\description{
Coefficients for the PVF LT, numerical deriv wrt. alpha
}

