% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{deriv_deriv_lt_dpvf_r_numeric}
\alias{deriv_deriv_lt_dpvf_r_numeric}
\title{PVF LT 2nd deriv wrt. alpha, evaluated numerically}
\usage{
deriv_deriv_lt_dpvf_r_numeric(p, s, alpha)
}
\description{
PVF LT 2nd deriv wrt. alpha, evaluated numerically
}

