% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{deriv_dgamma_r_numeric}
\alias{deriv_dgamma_r_numeric}
\title{Deriv of gamma density wrt. theta, evaluated numerically}
\usage{
deriv_dgamma_r_numeric(x, theta)
}
\description{
Deriv of gamma density wrt. theta, evaluated numerically
}

