% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_dposstab_r_numeric}
\alias{deriv_dposstab_r_numeric}
\title{Since there's no closed-form expression for the posstab density, the
derivative is evaluated numerically. This isn't needed for estimation}
\usage{
deriv_dposstab_r_numeric(x, alpha)
}
\description{
Since there's no closed-form expression for the posstab density, the
derivative is evaluated numerically. This isn't needed for estimation
}

