% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{deriv_lt_dgamma_r}
\alias{deriv_lt_dgamma_r}
\title{Derivative of the Laplace transform of the gamma density, wrt theta,
pth derivative wrt. s}
\usage{
deriv_lt_dgamma_r(p, s, theta)
}
\description{
Derivative of the Laplace transform of the gamma density, wrt theta,
pth derivative wrt. s
}

