% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_lt_dposstab_r}
\alias{deriv_lt_dposstab_r}
\title{Derivative wrt. alpha of the posstab LT.
TODO: determine coefficients for the pth derivatev wrt. s
currently only p = 0 (0th derivative wrt. s) has a closed-form expression}
\usage{
deriv_lt_dposstab_r(p, s, alpha)
}
\description{
Derivative wrt. alpha of the posstab LT.
TODO: determine coefficients for the pth derivatev wrt. s
currently only p = 0 (0th derivative wrt. s) has a closed-form expression
}

