% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_lt_dposstab_r_numeric}
\alias{deriv_lt_dposstab_r_numeric}
\title{Derivative wrt. alpha of the posstab LT}
\usage{
deriv_lt_dposstab_r_numeric(p, s, alpha)
}
\description{
Derivative wrt. alpha of the posstab LT
}

