% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_lt_dpvf_r}
\alias{deriv_lt_dpvf_r}
\title{Similar issue as deriv_lt_dposstab_r, only closed-form expression for the
0th derivative wrt. s right now}
\usage{
deriv_lt_dpvf_r(p, s, alpha)
}
\description{
Similar issue as deriv_lt_dposstab_r, only closed-form expression for the
0th derivative wrt. s right now
}

