% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{deriv_lt_dpvf_r_numeric}
\alias{deriv_lt_dpvf_r_numeric}
\title{Derivative wrt. alpha of the PVF LT, evaluated numerically}
\usage{
deriv_lt_dpvf_r_numeric(p, s, alpha)
}
\description{
Derivative wrt. alpha of the PVF LT, evaluated numerically
}

