% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{dlognormalmix_r}
\alias{dlognormalmix_r}
\title{A bimodal log-normal mixture, given by the variance and distance between means}
\usage{
dlognormalmix_r(x, theta)
}
\description{
A bimodal log-normal mixture, given by the variance and distance between means
}

