% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{dposstab_r}
\alias{dposstab_r}
\title{Positive stable density
TODO: does not converge when x~0 for small alpha. The point where convergance
fails depends on alpha
eg. with alpha = 0.5, converges at x = 0.05, but not at 0.01}
\usage{
dposstab_r(x, alpha, K = 100)
}
\description{
Positive stable density
TODO: does not converge when x~0 for small alpha. The point where convergance
fails depends on alpha
eg. with alpha = 0.5, converges at x = 0.05, but not at 0.01
}

