% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpvf_r}
\alias{dpvf_r}
\title{See comments from dposstab_r, same convergence issue applies here}
\usage{
dpvf_r(x, alpha, K = 100)
}
\description{
See comments from dposstab_r, same convergence issue applies here
}

