% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitfrail.control.R
\name{fitfrail.control}
\alias{fitfrail.control}
\title{Initialize control parameters}
\usage{
fitfrail.control(fitmethod = "score", abstol = 1e-08, reltol = 1e-06,
  iter.max = 100, verbose = FALSE)
}
\description{
Initialize control parameters
}

