\name{frailtyr-package}
\alias{frailtyr-package}
\alias{frailtyr}
\docType{package}
\title{
\packageTitle{frailtyr}
}
\description{
\packageDescription{frailtyr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{frailtyr}
\packageIndices{frailtyr}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{frailtyr}

Maintainer: \packageMaintainer{frailtyr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
