% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailtyr.R
\docType{package}
\name{frailtyr}
\alias{frailtyr}
\alias{frailtyr-package}
\title{frailtyr: A package for general semiparametric shared frailty models}
\description{
The frailtyr package provides:
foo, bar and baz.
}
\section{frailtyr functions}{

The functions ...
}

