% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genfrail.R
\name{genfrail}
\alias{genfrail}
\title{Generate shared frailty survival data}
\usage{
genfrail(...)
}
\arguments{
\item{...}{model}
}
\value{
parameters
  \url{} for more details.
}
\description{
\code{genfrail} generates data.
}
\details{
This function ...
}
\examples{
genfrail()
}

