% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{lt_deriv_dpvf_coef_r_numeric}
\alias{lt_deriv_dpvf_coef_r_numeric}
\title{Coefficients for the PVF LT, numerical deriv wrt. alpha}
\usage{
lt_deriv_dpvf_coef_r_numeric(p, j, alpha)
}
\description{
Coefficients for the PVF LT, numerical deriv wrt. alpha
}

