% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{lt_dgamma_moment_r}
\alias{lt_dgamma_moment_r}
\title{pth moment of the gamma distribution, evaluated by the gamma density LT}
\usage{
lt_dgamma_moment_r(p, theta)
}
\description{
pth moment of the gamma distribution, evaluated by the gamma density LT
}

