% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailtydists.R
\name{lt_dgamma_r}
\alias{lt_dgamma_r}
\title{Laplace transform of the gamma density, as defined above}
\usage{
lt_dgamma_r(s, theta, p)
}
\description{
Laplace transform of the gamma density, as defined above
}

