% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{lt_dposstab_moment_r}
\alias{lt_dposstab_moment_r}
\title{pth moment of the posstab distribution, evaluated using the LT}
\usage{
lt_dposstab_moment_r(alpha, p)
}
\description{
pth moment of the posstab distribution, evaluated using the LT
}

