% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{lt_dposstab_r}
\alias{lt_dposstab_r}
\title{Laplace transform of the positive stable distribution}
\usage{
lt_dposstab_r(p, s, alpha)
}
\description{
Laplace transform of the positive stable distribution
}

