% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{lt_dpvf_moment_r}
\alias{lt_dpvf_moment_r}
\title{pth moment of the one-parameter PVF distribution, as defined above}
\usage{
lt_dpvf_moment_r(alpha, p)
}
\description{
pth moment of the one-parameter PVF distribution, as defined above
}

