% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{lt_dpvf_r}
\alias{lt_dpvf_r}
\title{Laplace transform of the one-parameter PVF distribution defined above}
\usage{
lt_dpvf_r(p, s, alpha)
}
\description{
Laplace transform of the one-parameter PVF distribution defined above
}

