% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailtydists.R
\name{phi_dgamma}
\alias{phi_dgamma}
\title{Example of calculating phi_1 with numerical integrals}
\usage{
phi_dgamma(H_dot, N_dot, theta, k = 1)
}
\description{
Example of calculating phi_1 with numerical integrals
}

