% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{phi_laplace}
\alias{phi_laplace}
\title{Example of calculating phi_ with the LT}
\usage{
phi_laplace(k, N_dot, H_dot, density_params, density_LT)
}
\description{
Example of calculating phi_ with the LT
}

