% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailty.R
\name{phi_numerical}
\alias{phi_numerical}
\title{Example of calculating phi_ with numerical integrals}
\usage{
phi_numerical(k, N_dot, H_dot, density_params, density_fn)
}
\description{
Example of calculating phi_ with numerical integrals
}

