% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.summary.simfrail.R
\name{print.summary.simfrail}
\alias{print.summary.simfrail}
\title{Print a summary of the simulation}
\usage{
\method{print}{summary.simfrail}(sum.sim, n.Lambda = 3,
  digits = max(options()$digits - 3, 3), ...)
}
\description{
Print a summary of the simulation
}

