% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlaptrans.R
\name{rlaptrans}
\alias{rlaptrans}
\title{rlaptrans taken from Martin Ridout:
Ridout, M.S. (2009) Generating random numbers from a distribution specified
by its Laplace transform. Statistics and Computing, 19, 439-450.
http://www.kent.ac.uk/smsas/personal/msr/rlaptrans.html}
\usage{
rlaptrans(n, ltpdf, ..., tol = 1e-07, x0 = 1, xinc = 2, m = 11, L = 1,
  A = 19, nburn = 38)
}
\description{
rlaptrans taken from Martin Ridout:
Ridout, M.S. (2009) Generating random numbers from a distribution specified
by its Laplace transform. Statistics and Computing, 19, 439-450.
http://www.kent.ac.uk/smsas/personal/msr/rlaptrans.html
}

