% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frailtydists.R
\name{rposstab}
\alias{rposstab}
\title{Random generation from a positive stable distribution}
\usage{
rposstab(n, alpha)
}
\arguments{
\item{n}{number of samples to generate}

\item{alpha}{index of the distribution, where 0 < alpha < 1}
}
\value{
w
}
\description{
\code{rposstab} generates data.
}
\details{
Based on Chambers
}

