% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simfrail.R
\name{simfrail}
\alias{simfrail}
\title{Repeatedly simulate survival data and fit a model}
\usage{
simfrail(reps, genfrail.args, fitfrail.args, Lambda.time, vcov.args = list(),
  cores = 0, verbose = FALSE, skip.SE = FALSE)
}
\arguments{
\item{reps}{number of times to repeat the simulation}

\item{genfrail.args}{}

\item{fitfrail.args}{}

\item{Lambda.time}{}

\item{cores}{}
}
\value{
sim
}
\description{
Generate data and fit a model many times.
Reproduceable results can be obtained by simply setting the seed before calling simfrail.
}

