% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{theta.given.tau}
\alias{theta.given.tau}
\title{Numericcaly determine theta given Kendall's tau}
\usage{
theta.given.tau(tau, frailty)
}
\description{
Numericcaly determine theta given Kendall's tau
}

