% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcov.fitfrail.R
\name{vcov.fitfrail}
\alias{vcov.fitfrail}
\title{Compute the variance/covariance matrix for a fitted object}
\usage{
\method{vcov}{fitfrail}(fit, boot = TRUE, B = 100)
}
\description{
Compute the variance/covariance matrix for a fitted object
}

