% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_deriv_lt_dgamma_r_numeric}
\alias{deriv_deriv_lt_dgamma_r_numeric}
\title{Gamma LT 2nd derivative wrt. theta, evaluated numerically}
\usage{
deriv_deriv_lt_dgamma_r_numeric(m, s, theta)
}
\description{
Gamma LT 2nd derivative wrt. theta, evaluated numerically
}

