% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{deriv_lt_dgamma_r_numeric}
\alias{deriv_lt_dgamma_r_numeric}
\title{Same as above, evaluated numerically}
\usage{
deriv_lt_dgamma_r_numeric(p, s, theta)
}
\description{
Same as above, evaluated numerically
}

