% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitfrail.control.R
\name{fitfrail.control}
\alias{fitfrail.control}
\title{Initialize control parameters}
\usage{
fitfrail.control(fitmethod = "score", eps = 1e-09,
  toler.chol = .Machine$double.eps^0.75, iter.max = 20,
  toler.inf = sqrt(eps), verbose = TRUE)
}
\description{
Initialize control parameters
}

