% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitfrail.R
\name{fitfrail}
\alias{fitfrail}
\title{Fit a Shared Frailty model}
\usage{
fitfrail(data)
}
\arguments{
\item{...}{model}
}
\value{
parameters
  \url{} for more details.
}
\description{
\code{fitfrail} fits a model.
}
\details{
This function ...
}
\examples{
fitfrail(genfrail())
}

