% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genfrail.R
\name{genfrail}
\alias{genfrail}
\title{Generate shared frailty survival data}
\usage{
genfrail(beta = c(log(2)), frailty = c("gamma", "lognormal", "posstab"),
  censor.mu = 130, censor.sigma = 15, theta = 2, N = 300, K = 2,
  tau = 4.6, C = 0.01, covariates = c("normal", "uniform"))
}
\arguments{
\item{...}{model}
}
\value{
parameters
  \url{} for more details.
}
\description{
\code{genfrail} generates data.
}
\details{
This function ...
}
\examples{
genfrail()
}

