% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{rposstab_r}
\alias{rposstab_r}
\title{Random generation from a positive stable distribution}
\usage{
rposstab_r(n, alpha)
}
\description{
Random generation from a positive stable distribution
}

